requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
numpy>=1.24.0